+++ SQL_commands.sql
@@ -3,12 +3,12 @@ CREATE DATABASE IF NOT EXISTS gzauto;
 USE gzauto;
 
 CREATE TABLE Account(
-	Id INT PRIMARY KEY,
+	Id INT PRIMARY KEY AUTO_INCREMENT,
     Username VARCHAR(25) NOT NULL,
     Password VARCHAR(80) NOT NULL
 );
 CREATE TABLE Information(
-	Id INT PRIMARY KEY,
+	Id INT PRIMARY KEY AUTO_INCREMENT,
     FName VARCHAR(20) NOT NULL,
     MName VARCHAR(40),
     LName VARCHAR(30) NOT NULL,
@@ -22,7 +22,20 @@ CREATE TABLE Information(
     Account_id INT,
     CONSTRAINT FOREIGN KEY (Account_id) REFERENCES Account(Id)
 );
+CREATE TABLE Employees(
+	Id INT PRIMARY KEY AUTO_INCREMENT,
+	Username VARCHAR(25) NOT NULL,
+    Password VARCHAR(80) NOT NULL,
+    Role VARCHAR(15) NOT NULL,
+    FName VARCHAR(20) NOT NULL,
+    MName VARCHAR(40),
+    LName VARCHAR(30) NOT NULL,
+    Email VARCHAR(50),
+    Cellphone VARCHAR(16),
+    Birthday DATETIME
+);
 SHOW TABLES;
 DESC account;
 DESC information;
-DROP TABLE information;
+DROP TABLE Account;
+DROP TABLE Information;