+++ app/models.py
@@ -1,4 +1,4 @@
-from app import db
+from app import db, login_manager
 from flask_login import UserMixin
 from app import Bcrypt
 
@@ -8,14 +8,14 @@ Criando as tabelas do nosso banco de dados
 class Account(db.Model, UserMixin):
     Id = db.Column(db.Integer(), primary_key=True)
     Username = db.Column(db.String(length=25), nullable=False, unique=True)
-    Password = db.Column(db.String(length=80))
+    Password = db.Column(db.String(length=80), nullable=False)
     @property
     def password(self):
-        return self.password
+        return self.Password
     
     @password.setter
-    def password_hash(self, password):
-        self.password = Bcrypt.generate_password_hash(password).decode('utf-8')
+    def password_hash(self, Password):
+        self.Password = Bcrypt.generate_password_hash(Password).decode('utf-8')
 
     def check_password(self, attempted_password):
         return Bcrypt.check_password_hash(self.Password, attempted_password)
@@ -42,7 +42,36 @@ class Information(db.Model, UserMixin):
     #account = db.relationship('Account', backref=db.backref('information', uselist=False), lazy=True)
 
 class Employees(db.Model, UserMixin):
-    IdEmployees = db.Column(db.Integer(), primary_key=True)
+    Id = db.Column(db.Integer(), primary_key=True)
+    Username = db.Column(db.String(length=25), nullable=False)
+    Password = db.Column(db.String(length=80), nullable=False)
     FName = db.Column(db.String(length=20), nullable=False)
     MName = db.Column(db.String(length=40))
     LName = db.Column(db.String(length=30), nullable=False)
+    Email = db.Column(db.String(length=50))
+    Cellphone = db.column(db.String(length=16))
+    Birthday = db.Column(db.DateTime())
+    def get_id(self):
+        return int(self.Id)
+    
+    @property
+    def Password(self):
+        return self.Password
+    
+    @Password.setter
+    def password_hash(self, password):
+        self.password = Bcrypt.generate_password_hash(password).decode('utf-8')
+
+    def check_password(self, attempted_password):
+        return Bcrypt.check_password_hash(self.Password, attempted_password)
+
+@login_manager.user_loader
+def load_user(user):
+    user_type, id = user.split(":")
+
+    if user_type == "account":
+        return Account.query.get(id)
+    elif user_type == "employee":
+        return Employees.query.get(id)
+    else:
+        return None